Removed unnecessary static-locals.
In this file there's no access to the W5100 auto-increment registers so there's no need for the special compiler settings. However, with "true" local variables the buffer to read the HTTP header into is better not placed in a local variable.

Note: The heap manager was already active before as both linenoise and the file I/O use it.